+++ src/control_plane/project_infrastructure/mod.rs
@@ -1,4 +1,5 @@
 mod plan_elasticsearch_project_resources;
+mod plan_localstack_project_resources;
 mod plan_meilisearch_project_resources;
 mod plan_memcached_project_resources;
 mod plan_opensearch_project_resources;
@@ -10,6 +11,7 @@ mod project_service_preparation_error;
 mod project_service_preparation_strategy;
 
 pub(crate) use plan_elasticsearch_project_resources::plan_elasticsearch_project_resources;
+pub(crate) use plan_localstack_project_resources::plan_localstack_project_resources;
 pub(crate) use plan_meilisearch_project_resources::plan_meilisearch_project_resources;
 pub(crate) use plan_memcached_project_resources::plan_memcached_project_resources;
 pub(crate) use plan_opensearch_project_resources::{